moved pagehtml and thread into a specific folder + removed linting warning
