+++ src/container/WorkspaceContent.jsx
@@ -3,8 +3,8 @@ import { connect } from 'react-redux'
 import Folder from '../component/Workspace/Folder.jsx'
 import FileItem from '../component/Workspace/FileItem.jsx'
 import FileItemHeader from '../component/Workspace/FileItemHeader.jsx'
-// import Thread from './Thread.jsx'
-import PageText from './PageText.jsx'
+// import Thread from './FileType/Thread.jsx'
+import PageHtml from './FileType/PageHtml.jsx'
 import PageWrapper from '../component/common/layout/PageWrapper.jsx'
 import PageTitle from '../component/common/layout/PageTitle.jsx'
 import PageContent from '../component/common/layout/PageContent.jsx'
@@ -57,7 +57,7 @@ class WorkspaceContent extends React.Component {
 
           <DropdownCreateButton customClass='workspace__content__button mb-5' />
 
-          <PageText visible={this.state.activeFileType === 'file'} />
+          <PageHtml visible={this.state.activeFileType === 'file'} />
           {/*
           <Thread visible={this.state.activeFileType === 'chat'} />
           */}
