+++ src/css/index.styl
@@ -18,5 +18,5 @@ html, body, #content, #content > div
 @import 'Folder'
 
 @import 'Thread'
-@import 'PageText'
+@import 'PageHtml'
 @import 'Timeline'
